backend/requirements.txt
Flask
openai
python-dotenv
Flask-CORS
gunicorn
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended